🔨 Set biglist items to always be bunched together
If someone tried viewing the page with some sort of ultrawide monitor,
the `justify-content = space-around` would result in the different items
of a biglist being really spread apart. Changing this to `justify-content = center`
prevents this from happening because the items all stay close to the
center even if there is a bunch of empty space on the sides of the
biglist.